requirements.txt
aiofiles==23.2.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
et-xmlfile==1.1.0
exceptiongroup==1.2.1
fastapi==0.110.2
ffmpy==0.3.2
filelock==3.13.4
fonttools==4.51.0
fsspec==2024.3.1
gradio==4.28.0
gradio_client==0.16.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.7
importlib_resources==6.4.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
MouseInfo==0.1.3
numpy==1.26.4
openpyxl==3.0.10
orjson==3.10.1
packaging==24.0
pandas==2.2.2
pillow==10.3.0
PyAutoGUI==0.9.54
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
PyGetWindow==0.0.9
Pygments==2.17.2
PyMsgBox==1.0.9
pyparsing==3.1.2
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytweening==1.2.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
ruff==0.4.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tomlkit==0.12.0
toolz==0.12.1
tqdm==4.66.2
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
Unidecode @ file:///home/conda/feedstock_root/build_artifacts/unidecode_1704986336983/work
urllib3==2.2.1
uvicorn==0.29.0
websockets==11.0.3
